requirements.txt
pytest
allure-pytest
pydentic
requests